feat : BulletinBoard MyPost 구현_1

+++ data/board/src/main/java/com/umcspot/spot/board/datasource/BoardDataSource.kt
@@ -15,7 +15,7 @@ interface BoardDataSource {
     ): BaseResponse<BestBoardResponseDto>
 
     suspend fun getFilteredPosts(
-        cursor : Int,
+        cursor : Long,
         postType: PostType,
         size : Int
     ): BaseResponse<PostResponseDto>
