+++ data/board/src/main/java/com/umcspot/spot/board/datasourceimpl/BoardDataSourceImpl.kt
@@ -23,7 +23,7 @@ class BoardDataSourceImpl @Inject constructor(
         boardService.getBestBoard(sortBy)
 
     override suspend fun getFilteredPosts(
-        cursor : Int,
+        cursor : Long,
         postType: PostType,
         size : Int
     ): BaseResponse<PostResponseDto> =
