+++ domain/board/src/main/java/com/umcspot/spot/domain/board/repository/BoardRepository.kt
@@ -2,7 +2,7 @@ package com.umcspot.spot.domain.board.repository
 
 import com.umcspot.spot.domain.board.model.board.BestPostResultList
 import com.umcspot.spot.domain.board.model.board.RecentPostResultList
-import com.umcspot.spot.domain.board.model.post.PostResultList
+import com.umcspot.spot.domain.board.model.postList.PostResultList
 import com.umcspot.spot.model.PostType
 import com.umcspot.spot.model.SortType
 
@@ -12,5 +12,5 @@ interface BoardRepository {
 
     suspend fun getBestBoard(sortBy: SortType): Result<BestPostResultList>
 
-    suspend fun getFilteredPosts(cursor : Int? = null, postType: PostType? = null, size : Int): Result<PostResultList>
+    suspend fun getFilteredPosts(cursor : Long? = null, postType: PostType? = null, size : Int): Result<PostResultList>
 }
